# improve performance of successful int extract by ~30% by avoiding calls to `__index__` where redundant