--- a/newsfragments/3742.changed.md
+++ b/newsfragments/3742.changed.md
@@ -1,0 +1,1 @@
+Improve performance of `extract::<i64>` (and other integer types) by avoiding call to `__index__()` converting the value to an integer for 3.10+. Gives performance improvement of around 30% for successful extraction.