@davidhewitt 2024-01-14
https://github.com/python/cpython/issues/80229 looks relevant,  which was in 3.8 so we may need to keep a different path for 3.7.