@samuelcolvin 2024-01-14
Logic changed for 3.8 to just call `PyLong_AsLong`, with this the performance increased to 37% improvement.

I haven't changed the extraction logic for `BigInt` since that is ultimately calling `_PyLong_NumBits` which doesn't internally call `PyNumber_Index`, so I think it's right - I guess it's also used much less.