@davidhewitt 2024-01-15
Uff unfortunately the test failure looks legitimate, `PyLong_AsUnsignedLong` and `PyLong_AsUnsignedLongLong` do _not_ call `index`, so we will have to manually wrap those to call `__index__` again. 

https://github.com/python/cpython/blob/41a94c9e7be94760baab1dcb33427d8781bea64a/Objects/longobject.c#L630C68-L630C68