@samuelcolvin 2024-01-15
turns out the benchmarks extract a lot of `usize`s...