@samuelcolvin 2024-01-16
Looks like this will fail, I think the lossy float conversion wasn't fixed until 3.10 in https://github.com/python/cpython/issues/82180.