@davidhewitt 2024-01-16
Ok great; so now we can do the fast-path on 3.10+ instead. Makes sense.