@samuelcolvin 2024-01-16
I'm assuming the segmentation fault with numpy on the merge queue is not something for me to fix?